Fix for masonry tests
